+++ lib/main.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+import 'views/lista_transferencia.dart';
+
 void main() => runApp(BytebankApp());
 
 class BytebankApp extends StatelessWidget {
@@ -15,157 +17,5 @@ class BytebankApp extends StatelessWidget {
       ),
       home: ListaTransferencias(),
     );
-  }
-}
-
-class FormularioTransferencia extends StatefulWidget {
-  final TextEditingController _controladorCampoNumeroConta =
-      TextEditingController();
-  final TextEditingController _controladorCampoValor = TextEditingController();
-
-  @override
-  State<StatefulWidget> createState() {
-    return FormularioTransferenciaState();
-  }
-}
-
-class FormularioTransferenciaState extends State<FormularioTransferencia> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Criando Transferência'),
-      ),
-      body: SingleChildScrollView(
-        child: Column(
-          children: [
-            Editor(
-              rotulo: 'Número da conta',
-              dica: '0000',
-              controlador: widget._controladorCampoNumeroConta,
-            ),
-            Editor(
-              rotulo: 'Valor',
-              dica: '0.00',
-              controlador: widget._controladorCampoValor,
-              icone: Icon(Icons.monetization_on),
-            ),
-            RaisedButton(
-              child: Text('Confirmar'),
-              onPressed: () => _criaTransferencia(context),
-            )
-          ],
-        ),
-      ),
-    );
-  }
-
-  void _criaTransferencia(BuildContext context) {
-    final int numeroConta =
-        int.tryParse(widget._controladorCampoNumeroConta.text);
-    final double valor = double.tryParse(widget._controladorCampoValor.text);
-    if (numeroConta != null && valor != null) {
-      final transferencia = Transferencia(valor, numeroConta);
-      Navigator.pop(context, transferencia);
-    }
-  }
-}
-
-class Editor extends StatelessWidget {
-  final TextEditingController controlador;
-  final String rotulo;
-  final String dica;
-  final Icon icone;
-
-  Editor({this.rotulo, this.dica, this.icone, this.controlador});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      child: TextField(
-        controller: controlador,
-        style: TextStyle(
-          fontSize: 20.0,
-        ),
-        decoration: InputDecoration(
-          icon: icone,
-          labelText: rotulo,
-          hintText: dica,
-        ),
-        keyboardType: TextInputType.number,
-      ),
-      padding: EdgeInsets.all(16.0),
-    );
-  }
-}
-
-class ListaTransferencias extends StatefulWidget {
-  final List<Transferencia> _transferencias = List();
-
-  @override
-  State<StatefulWidget> createState() {
-    return ListaTransferenciasState();
-  }
-}
-
-class ListaTransferenciasState extends State<ListaTransferencias> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Transferências'),
-      ),
-      body: ListView.builder(
-        itemBuilder: (context, indice) {
-          final Transferencia transferencia = widget._transferencias[indice];
-          return ItemTransferencia(transferencia);
-        },
-        itemCount: widget._transferencias.length,
-      ),
-      floatingActionButton: FloatingActionButton(
-        child: Icon(Icons.add),
-        onPressed: () {
-          final Future<Transferencia> future =
-              Navigator.push(context, MaterialPageRoute(builder: (context) {
-            return FormularioTransferencia();
-          }));
-          future.then((transferenciaCriada) {
-            debugPrint('$transferenciaCriada');
-            if (transferenciaCriada != null) {
-              setState(() => widget._transferencias.add(transferenciaCriada));
-            }
-          });
-        },
-      ),
-    );
-  }
-}
-
-class ItemTransferencia extends StatelessWidget {
-  final Transferencia _transferencia;
-
-  ItemTransferencia(this._transferencia);
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      child: ListTile(
-        leading: Icon(Icons.monetization_on),
-        title: Text(_transferencia.valor.toString()),
-        subtitle: Text(_transferencia.numeroConta.toString()),
-      ),
-    );
-  }
-}
-
-class Transferencia {
-  final double valor;
-  final int numeroConta;
-
-  Transferencia(this.valor, this.numeroConta);
-
-  @override
-  String toString() {
-    return 'valor: $valor - numero conta: $numeroConta';
   }
 }
