+++ lib/models/transferencia.dart
@@ -1,0 +1,11 @@
+class Transferencia {
+  final double valor;
+  final int numeroConta;
+
+  Transferencia(this.valor, this.numeroConta);
+
+  @override
+  String toString() {
+    return 'valor: $valor - numero conta: $numeroConta';
+  }
+}
